Move lib and bin into packages directory
