+++ packages/api-pls-cli/commands/start.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const log = require('../util/log');
-const startServer = require('../../lib/start-server');
+const startServer = require('../../api-pls-utils/start-server');
 
 module.exports = function(options) {
   log.info(`Starting the API webserver on port ${options.port}`);
