+++ packages/api-pls-express-router/index.js
@@ -12,7 +12,7 @@ const jsonApiHeaders = require('./util/json-api-headers');
 const adjustResourceQuantity = require('./util/adjust-resource-quantity');
 const jsonApiMediaType = require('./util/json-api-media-type');
 const log = require('./util/log');
-const generateDefinitions = require('../../lib/resource-definition/generate-from-raw');
+const generateDefinitions = require('../api-pls-utils/resource-definition/generate-from-raw');
 
 module.exports = function(options) {
   const router = express.Router();
