+++ packages/api-pls-postgres-adapter/middleware/create.js
@@ -4,7 +4,7 @@ const _ = require('lodash');
 const handleQueryError = require('../util/handle-query-error');
 const formatTransaction = require('../util/format-transaction');
 const crud = require('../sql/crud');
-const serverErrors = require('../../../lib/server-errors');
+const serverErrors = require('../../api-pls-utils/server-errors');
 
 module.exports = async function(req) {
   const pls = req.pls;
