+++ packages/api-pls-postgres-adapter/middleware/read.js
@@ -5,7 +5,7 @@ const qs = require('qs');
 const crud = require('../sql/crud');
 const handleQueryError = require('../util/handle-query-error');
 const formatTransaction = require('../util/format-transaction');
-const serverErrors = require('../../../lib/server-errors');
+const serverErrors = require('../../api-pls-utils/server-errors');
 
 module.exports = async function(req) {
   const pls = req.pls;
