+++ packages/api-pls-postgres-adapter/sql/sql-util.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const pgp = require('pg-promise');
-const relationshipUtil = require('../../../lib/relationship-util');
+const relationshipUtil = require('../../api-pls-utils/relationship-util');
 
 // Given a `resource`, returns the table name for it.
 // Pass `{escaped: true}` as options to escape the table name.
