+++ packages/api-pls-postgres-adapter/util/build-response-relationships.js
@@ -3,7 +3,7 @@
 const _ = require('lodash');
 const sqlUtil = require('../sql/sql-util');
 const manyToManyUtil = require('../sql/many-to-many-util');
-const relationshipUtil = require('../../../lib/relationship-util');
+const relationshipUtil = require('../../api-pls-utils/relationship-util');
 
 function formatToOneResult({result, definition, value, version, columnBase, relation, adjustResourceQuantity}) {
   const id = value ? String(value) : null;
