+++ packages/api-pls-utils/resource-definition/build-json-schema/util/require-id-validation.js
