+++ packages/api-pls-utils/resource-definition/generate-from-raw.js
