+++ packages/api-pls-utils/resource-model/load-from-disk.js
