+++ packages/api-pls-utils/resource-model/non-serializable-fields.json
