+++ packages/api-pls-utils/resource-model/normalize/util/normalize-fields.js
